script v2
some updates to the script for after the pilot. subj number now assigned, participants have to get the answers right before starting, and the subsequences are being noted, although I'm still working on having that row by row instead of the last overwriting all the previous ones. also I changed mouseover to mousemove so that now it also counts as a hit if your mouse was already positioned correctly.